+++ common/params.cc
@@ -85,6 +85,7 @@ private:
 
 std::unordered_map<std::string, uint32_t> keys = {
     // AleSato stuff 
+    {"AleSato_AutomaticBrakeHold", PERSISTENT},
     {"AleSato_RemoteLockDoors", CLEAR_ON_MANAGER_START | CLEAR_ON_ONROAD_TRANSITION},
     {"AleSato_ShutdownScreen", PERSISTENT},
     {"AleSato_SteerAlwaysOn", CLEAR_ON_MANAGER_START | CLEAR_ON_ONROAD_TRANSITION},