+++ selfdrive/car/toyota/toyotacan.py
@@ -110,3 +110,38 @@ def create_ui_command(packer, steer, chime, left_line, right_line, left_lane_dep
     ]})
 
   return packer.make_can_msg("LKAS_HUD", 0, values)
+
+
+def create_brakehold_command(packer, stock_AEB, cut3frames):
+  values = {
+    "DSS1GDRV": 1023,
+    "PBRTRGR": cut3frames,
+  }
+
+  if len(stock_AEB):
+    values.update({s: stock_AEB[s] for s in [
+      # They are not all necessary since brakehold only sends stopped,
+      # however we will leave prepared for the future alternative AEB of the comma.ai
+      "DSS1GDRV",
+      "DS1STAT2",
+      "DS1STBK2",
+      "PCSWAR",
+      "PCSALM",
+      "PCSOPR",
+      "PCSABK",
+      "PBATRGR",
+      "PPTRGR",
+      "IBTRGR",
+      "CLEXTRGR",
+      "IRLT_REQ",
+      "BRKHLD",
+      "AVSTRGR",
+      "VGRSTRGR",
+      "PREFILL",
+      "PBRTRGR",
+      "PCSDIS",
+      "PBPREPMP",
+    ]})
+
+  
+  return packer.make_can_msg("PRE_COLLISION_2", 0, values)