+++ RELEASES.md
@@ -5,6 +5,7 @@ Version 0.9.5 (202X-XX-XX)
 * Kia Sorento Hybrid 2023 support thanks to sunnyhaibin!
 * Lexus IS 2023 support thanks to L3R5!
 * Ale Sato
+  * Automatic Brake Hold (2023-09-18)
   * Lateral Tunning (fix hugg turn cutting) using Pfeiferj's distance based curvature and increased actuator delay (2023-08-10)
   * Fixed "controls lagging" bug when a engage is requested immediately after activating cruise main (2023-08-02)